+++ docs/public-api/crawlee-core.api.md
@@ -51,7 +51,6 @@ import type { StandardSchemaV1 } from '@standard-schema/spec';
 import type * as storage from '@crawlee/types';
 import { StorageBackend } from '@crawlee/types';
 import { StorageIdentifier } from '@crawlee/types';
-import { tryAbsoluteURL } from '@crawlee/utils';
 import { z } from 'zod';
 
 // @public (undocumented)
@@ -1762,8 +1761,6 @@ export interface SystemInfo {
     // (undocumented)
     memTotalBytes?: number;
 }
-
-export { tryAbsoluteURL }
 
 // @public (undocumented)
 export type UrlPatternObject = {
