+++ docs/public-api/crawlee-http.api.md
@@ -11,7 +11,7 @@ import { BasePredicate } from 'ow';
 import { BasicCrawler } from '@crawlee/basic';
 import { BasicCrawlerOptions } from '@crawlee/basic';
 import { BooleanPredicate } from 'ow';
-import { CheerioRoot } from '@crawlee/utils';
+import type { CheerioRoot } from '@crawlee/utils';
 import { ConcurrencySystem } from '@crawlee/basic';
 import type { ConcurrencySystemOptions } from '@crawlee/basic';
 import { ContextPipeline } from '@crawlee/basic';
