+++ packages/utils/src/internals/iterables.ts
@@ -1,5 +1,7 @@
 import { inspect } from 'node:util';
 
+import { asyncifyIterable } from './iterables-internal.js';
+
 /**
  * Type guard that checks if a value is iterable (has Symbol.iterator).
  * @internal
@@ -44,24 +46,6 @@ export function isAsyncIterable<T>(value: unknown): value is AsyncIterable<T> {
     }
 
     return typeof Object(value)[Symbol.asyncIterator] === 'function';
-}
-
-/**
- * Converts any iterable or async iterable to an async iterable.
- * @internal
- *
- * @yields Each item from the input iterable
- *
- * **Example usage:**
- * ```ts
- * const syncArray = [1, 2, 3];
- * for await (const item of asyncifyIterable(syncArray)) {
- *   console.log(item); // 1, 2, 3
- * }
- * ```
- */
-export async function* asyncifyIterable<T>(iterable: Iterable<T> | AsyncIterable<T>): AsyncIterable<T> {
-    yield* iterable;
 }
 
 /**
@@ -225,57 +209,4 @@ export function peekableAsyncIterable<T>(iterable: AsyncIterable<T> | Iterable<T
             return peekableIterator;
         },
     };
-}
-
-// Source - https://stackoverflow.com/a/71288323
-/**
- * Merges multiple async iterables into a single async iterable, yielding values concurrently.
- *
- * **Example usage:**
- * ```ts
- * const asyncIterable1 = async function* () {
- *   yield 1; yield 3; yield 5;
- * };
- *
- * const asyncIterable2 = async function* () {
- *   yield 2; yield 4; yield 6;
- * };
- *
- * for await (const value of mergeAsyncIterables(asyncIterable1(), asyncIterable2())) {
- *   console.log(value);
- * }
- * ```
- */
-export async function* mergeAsyncIterables<T>(...iterables: AsyncIterable<T>[]): AsyncIterable<T> {
-    const asyncIterators = iterables.map((iterable) => iterable[Symbol.asyncIterator]());
-    const results = [];
-    let count = asyncIterators.length;
-    const never: Promise<never> = new Promise(() => {});
-    async function getNext(asyncIterator: AsyncIterator<T>, index: number) {
-        const result = await asyncIterator.next();
-        return {
-            index,
-            result,
-        };
-    }
-    const nextPromises = asyncIterators.map(getNext);
-    try {
-        while (count) {
-            const { index, result } = await Promise.race(nextPromises);
-            if (result.done) {
-                nextPromises[index] = never;
-                results[index] = result.value;
-                count--;
-            } else {
-                nextPromises[index] = getNext(asyncIterators[index], index);
-                yield result.value;
-            }
-        }
-    } finally {
-        for (const [index, iterator] of asyncIterators.entries()) {
-            // no await here - see https://github.com/tc39/proposal-async-iteration/issues/126
-            if (nextPromises[index] !== never && iterator.return != null) void iterator.return();
-        }
-    }
-    return results;
 }
